# src/backend/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Car Share API
  description: This is the API for the Car Share project.
  version: 0.4.0 # TODO: Update version number
  contact:
    name: Car Share, backend team
    email: 'contoux1u@etu.univ-lorraine.fr'
  license:
    name: GNU General Public License v3.0
    url: 'https://www.gnu.org/licenses/gpl-3.0.en.html'
servers:
  - url: 'https://carshare.api.antoinectx.fr'
    description: Antoine's server
  - url: 'https://carshare.apis.furwaz.fr'
    description: Paul's server
...
paths:
...
  /travels/create:
    post:
      summary: Create a travel
      description: "Create a travel with steps.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: createTravel
      tags:
        - Travels
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxPassengers:
                  type: integer
                  example: 10
                price:
                  type: integer
                  example: 78
                description:
                  type: string
                  example: Voyage non fumeur
                groupId:
                  type: integer
                  nullable: true
                  example: 1
                steps:
                  type: array
                  items:
                    $ref: '#/components/schemas/StepInput'
                  example:
                    - label: 2ter Boulevard Charlemagne 54000 Nancy
                      city: Nancy
                      context: 54, Meurthe-et-Moselle, Grand Est
                      lat: 48.683188
                      lng: 6.1877
                      date: 2023-04-16T09:15:36.000Z
                    - label: 2ter Boulevard Charlemagne 54000 Laxou
                      city: Laxou
                      context: 54, Meurthe-et-Moselle, Grand Est
                      lat: 49.683188
                      lng: 6.5877
                      date: 2023-04-16T219:15:36.000Z
                    - label: 1801 chemin Napoléon 55000 Metz
                      city: Metz
                      context: 55, Moselle, Grand Est
                      lat: 78.683188
                      lng: 6.1877
                      date: 2023-04-17T09:15:36.000Z
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Travel created"
                  travel:
                    $ref: '#/components/schemas/PublicTravel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Date is required.
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/userNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
...
  /travels/search:
    get:
      summary: Search travels
      description: Get all the travels fitting with the given criteria
      operationId: searchTravels
      tags:
        - Travels
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: date
          in: query
          description: "Wanted departure date (at midnight, ex: in UTC+2 = 2023-03-15T22:00:00.000Z)"
          schema:
            type: string
            example: 2023-03-15T22:00:00.000Z
        - name: time
          in: query
          description: "Wanted departure time (only the time, in the format HH:mm, in the user's timezone)
          For example, if the user is in a timezone UTC+2, and wants to search for a travel starting at 10:00, he should send 10:00, not 08:00."
          schema:
            type: string
            required: false
            example: 10:00
        - name: startCity
          in: query
          description: Start city
          schema:
            type: string
            example: Nancy
        - name: startContext
          in: query
          description: Start context
          schema:
            type: string
            example: 54, Meurthe-et-Moselle, Grand Est
        - name: endCity
          in: query
          description: End city
          schema:
            type: string
            example: Metz
        - name: endContext
          in: query
          description: End context
          schema:
            type: string
            example: 55, Moselle, Grand Est
        - name: startLat
          in: query
          description: Start latitude
          schema:
            type: number
            example: 48.692054
        - name: startLng
          in: query
          description: Start longitude
          schema:
            type: number
            example: 6.184417
        - name: endLat
          in: query
          description: End latitude
          schema:
            type: number
            example: 49.119309
        - name: endLng
          in: query
          description: End longitude
          schema:
            type: number
            example: 6.175868
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The travels fitting the given criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelSearch'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Date is invalid.
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /travels/:id:
    get:
      summary: Get one travel
      description: Get one trips with the given id
      operationId: getTravel
      tags:
        - Travels
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: query
          description: Target travel id
          schema:
            type: number
            example: 1
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The travel with the given id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTravel'
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/travelNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Update travel (admin)
      description: "Update a travel with the given id. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: updateTravel
      tags:
        - Travels
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: Travel id
          required: true
          schema:
            type: integer
            example: 1
        - name: reason
          in: query
          description: Reason for the update
          schema:
            type: string
            example: Price changed
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Travel to update
        content:
          application/json:
            schema:
              type: object
              properties:
                maxPassengers:
                  type: integer
                  required: false
                  example: 10
                price:
                  type: integer
                  required: false
                  example: 78
                description:
                  type: string
                  required: false
                  example: Voyage non fumeur
                steps:
                  type: array
                  items:
                    $ref: '#/components/schemas/StepInput'
      responses:
        '200':
          description: Travel updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Travel updated"
                  travel:
                    $ref: '#/components/schemas/PublicTravel'
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/travelNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    delete:
      summary: Cancel travel (admin)
      description: "Cancel a travel with the given id. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: cancelTravel
      tags:
        - Travels
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: Travel id
          required: true
          schema:
            type: integer
        - name: reason
          in: query
          description: Reason for the cancellation
          schema:
            type: string
            example: Conducteur dangereux
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Travel cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Travel cancelled"
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/travelNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /travels:
    get:
      summary: All travels (admin)
      description: "Get all travels.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: getTravels
      tags:
        - Travels
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: All travels
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Travel'
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
                  next:
                    type: integer
                    nullable: true
                    example: 10
                  prev:
                    type: integer
                    nullable: true
                    example: 0
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    Accept-Language:
      name: Accept-Language
      in: header
      description: The language to use in the response message (if any)
      required: false
      schema:
        type: string
        default: en
        enum:
          - en
          - fr
    limit:
      name: limit
      in: query
      description: The number of items to return
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 50
    offset:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
...
    invalidId:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Id is invalid.
    noToken:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: No token provided.
    invalidToken:
      description: Invalid/Expired token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid token provided.
    insufficientPrivileges:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: You do not have sufficient rights to perform this action.
...
    userNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: User not found.
    travelNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Travel not found.
...
    internalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: An internal error occurred. If the problem persists, please contact an administrator.
...
  schemas:
...
    TravelSearch:
      type: object
      properties:
        id:
          type: integer
          example: 1
        maxPassengers:
          type: integer
          example: 4
        price:
          type: number
          example: 100
        description:
          type: string
          nullable: true
          example: No pets allowed
        status:
          type: integer
          example: 1
        driverId:
          type: integer
          example: 1
        groupId:
          type: integer
          nullable: true
          example: 1
        createdAt:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
        driver:
          $ref: '#/components/schemas/PublicUser'
        departure:
          $ref: '#/components/schemas/Step'
        arrival:
          $ref: '#/components/schemas/Step'
        passengers:
          type: integer
          example: 2
          description: The maximum number of passengers between the departure and the arrival
...
    PublicUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: john@example.com
        emailVerifiedOn:
          type: string
          format: date-time
          nullable: true
          default: null
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+330612345678"
        avatar:
          type: string
          nullable: true
          default: null
        gender:
          type: integer
          default: -1
        hasCar:
          type: boolean
          default: false
        level:
          type: integer
          default: 0
        lang:
          type: string
          nullable: true
          default: null
          example: fr
        timezone:
          type: string
          nullable: true
          default: null
          example: Europe/Paris
...
    Step:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        city:
          type: string
        context:
          type: string
        lat:
          type: number
        lng:
          type: number
        travelId:
          type: integer
        date:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
    StepInput:
      type: object
      properties:
        id:
          type: integer
          required: false
        label:
          type: string
        city:
          type: string
        context:
          type: string
        lat:
          type: number
        lng:
          type: number
        date:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
...
    Travel:
      type: object
      properties:
        id:
          type: integer
          example: 1
        maxPassengers:
          type: integer
          example: 4
          default: 1
        price:
          type: number
          example: 24.5
          default: 0
        description:
          type: string
          nullable: true
          example: "Non-smoking travel"
        driverId:
          type: integer
          example: 1
        groupId:
          type: integer
          nullable: true
          default: null
        createdAt:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
    PublicTravel:
      type: object
      properties:
        id:
          type: integer
          example: 1
        maxPassengers:
          type: integer
          example: 4
          default: 1
        price:
          type: number
          example: 24.5
          default: 0
        description:
          type: string
          nullable: true
          example: "Non-smoking travel"
        driverId:
          type: integer
          example: 1
        driver:
          $ref: '#/components/schemas/PublicUser'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        createdAt:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
...